sensors utility created from code of background service
